added EditText style, button style and loading data animation
