+++ .idea/modules.xml
@@ -2,6 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/RawrProject.iml" filepath="$PROJECT_DIR$/RawrProject.iml" />
       <module fileurl="file://$PROJECT_DIR$/RawrProyect.iml" filepath="$PROJECT_DIR$/RawrProyect.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
