+++ app/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:name=".mainActivities.Loading_window"
+            android:name=".mainActivities.Welcome_screen"
             android:configChanges="orientation"
             android:label="@string/app_name"
             android:screenOrientation="portrait" >
@@ -25,7 +25,7 @@
             android:screenOrientation="portrait" >
         </activity>
         <activity
-            android:name=".mainActivities.SingUp"
+            android:name=".mainActivities.SignUp"
             android:configChanges="orientation"
             android:label="@string/app_name"
             android:screenOrientation="portrait" >
