+++ app/src/main/java/com/example/david/rawr/mainActivities/Loading_screen.java
@@ -1,7 +1,6 @@
 package com.example.david.rawr.mainActivities;
 
 import android.app.Activity;
-import android.content.Intent;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -9,38 +8,25 @@ import android.widget.TextView;
 
 import com.example.david.rawr.R;
 
-import java.util.Timer;
-import java.util.TimerTask;
 
+public class Loading_screen extends Activity {
 
-public class Loading_window extends Activity {
-
-    int count = 0;
     TextView icon;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_loading_window);
-        Timer timer = new Timer();
-        timer.schedule(new TimerTask() {
-            @Override
-            public void run() {
-                Intent intent = new Intent(Loading_window.this, LogIn.class );
-                startActivity(intent);
-                finishscreen();
-            }
-        }, 3000);
+        setContentView(R.layout.activity_loading_screen);
     }
 
-    private void finishscreen(){
-        this.finish();
-    }
+    private void finishscreen() {
+                this.finish();
+            }
 
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_loading_window, menu);
+        getMenuInflater().inflate(R.menu.loading_screen, menu);
         return true;
     }
