+++ app/src/main/java/com/example/david/rawr/mainActivities/LogIn.java
@@ -2,20 +2,16 @@ package com.example.david.rawr.mainActivities;
 
 import android.app.Activity;
 import android.content.Intent;
-import android.os.AsyncTask;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
 
 import com.example.david.rawr.R;
-import com.example.david.rawr.db.DbMethods;
 import com.example.david.rawr.db.ValidateUser;
 
 import java.util.concurrent.ExecutionException;
@@ -49,7 +45,7 @@ public class LogIn extends Activity implements View.OnClickListener {
             case (R.id.signUp):
                 // TODO
                 Toast.makeText(this, "Sing Up", Toast.LENGTH_LONG).show();
-                intent = new Intent(LogIn.this, SingUp.class );
+                intent = new Intent(LogIn.this, SignUp.class );
                 startActivity(intent);
                 finish_screen();
                 break;
@@ -58,7 +54,7 @@ public class LogIn extends Activity implements View.OnClickListener {
                 try {
                     String status = validate.execute().get();
                     if(status.compareTo("1") == 0) {
-                        intent = new Intent(LogIn.this, downloading_window.class);
+                        intent = new Intent(LogIn.this, Loading_screen.class);
                         startActivity(intent);
                         Toast.makeText(this, "Logged in", Toast.LENGTH_LONG).show();
                         finish_screen();
