--- app/src/main/java/com/example/david/rawr/mainActivities/downloading_window.java
@@ -1,51 +1,0 @@
-package com.example.david.rawr.mainActivities;
-
-import android.app.Activity;
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.Menu;
-import android.view.MenuItem;
-import android.widget.TextView;
-import com.example.david.rawr.otherClasses.fontType;
-import com.example.david.rawr.R;
-
-import java.util.Timer;
-import java.util.TimerTask;
-
-
-public class downloading_window extends Activity {
-
-    TextView icon;
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_downloading_window);
-    }
-
-            private void finishscreen() {
-                this.finish();
-            }
-
-
-            @Override
-            public boolean onCreateOptionsMenu(Menu menu) {
-                // Inflate the menu; this adds items to the action bar if it is present.
-                getMenuInflater().inflate(R.menu.menu_downloading_window, menu);
-                return true;
-            }
-
-            @Override
-            public boolean onOptionsItemSelected(MenuItem item) {
-                // Handle action bar item clicks here. The action bar will
-                // automatically handle clicks on the Home/Up button, so long
-                // as you specify a parent activity in AndroidManifest.xml.
-                int id = item.getItemId();
-
-                //noinspection SimplifiableIfStatement
-                if (id == R.id.action_settings) {
-                    return true;
-                }
-
-                return super.onOptionsItemSelected(item);
-            }
-        }
