+++ app/src/main/res/drawable/edit_text_style.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item android:state_focused="false">
+        <shape>
+            <solid android:color="#DAFF1B"/>
+            <corners android:radius="10dp"/>
+            <padding
+                android:bottom="15dp"
+                android:left="5dp"
+                android:right="5dp"
+                android:top="15dp"
+                />
+        </shape>
+    </item>
+
+    <item android:state_focused="true">
+        <shape>
+            <solid android:color="#FAFF1B"/>
+            <corners android:radius="10dp"/>
+            <padding
+                android:bottom="15dp"
+                android:left="5dp"
+                android:right="5dp"
+                android:top="15dp"
+                />
+        </shape>
+    </item>
+
+</selector>
