+++ app/src/main/res/layout/activity_loading_screen.xml
