+++ app/src/main/res/layout/login_window.xml
@@ -36,6 +36,7 @@
                 android:hint="@string/user"
                 android:layout_gravity="center_horizontal"
                 android:gravity="center_horizontal"
+                android:background="@drawable/edit_text_style"
                 />
 
             <EditText
@@ -46,13 +47,17 @@
                 android:id="@+id/passText"
                 android:layout_gravity="center_horizontal"
                 android:hint="@string/pass"
-                android:gravity="center_horizontal" />
+                android:gravity="center_horizontal"
+                android:background="@drawable/edit_text_style"
+                />
 
             <Button
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:text="@string/iniciarSesion"
-                android:id="@+id/logInButton" />
+                android:id="@+id/logInButton"
+                android:background="@drawable/button_style"
+                />
 
             <TextView
                 android:layout_marginTop="5dp"
