+++ app/src/main/res/menu/loading_screen.xml
